// src/app/models/jobs.model.ts
export interface Jobs {
    id: number;
    title: string;
    type: string;
    assigned_to: string;
    phone: string;
    assigned_by: string;
    dueDate: string;
    date: string;
    status: string;

}
export interface Partnerjobs {
    id: number;
    title: string;
    type: string;
    assigned_to: string;
    phone: string;
    assigned_by: string;
    dueDate: string;
    date: string;
    status: string;

}